Move Action and Line into todo file module
