+++ src/process/testutil.rs
@@ -4,11 +4,11 @@ use crate::display::curses::{Curses, Input as PancursesInput};
 use crate::display::Display;
 use crate::input::input_handler::InputHandler;
 use crate::input::Input;
-use crate::list::line::Line;
 use crate::process::exit_status::ExitStatus;
 use crate::process::process_module::ProcessModule;
 use crate::process::process_result::ProcessResult;
 use crate::process::state::State;
+use crate::todo_file::line::Line;
 use crate::todo_file::TodoFile;
 use crate::view::testutil::render_view_data;
 use crate::view::view_data::ViewData;
